refactor: remove gurl_converter from native_mate
