+++ shell/browser/web_contents_permission_helper.cc
@@ -12,7 +12,6 @@
 #include "content/public/browser/render_process_host.h"
 #include "shell/browser/atom_permission_manager.h"
 #include "shell/browser/media/media_stream_devices_controller.h"
-#include "shell/common/native_mate_converters/gurl_converter.h"
 
 namespace {
