+++ shell/browser/web_view_guest_delegate.cc
@@ -14,7 +14,6 @@
 #include "content/public/browser/render_widget_host.h"
 #include "content/public/browser/render_widget_host_view.h"
 #include "shell/browser/api/atom_api_web_contents.h"
-#include "shell/common/native_mate_converters/gurl_converter.h"
 #include "third_party/blink/public/common/page/page_zoom.h"
 
 namespace electron {
